+++ src/types.rs
@@ -1,4 +1,5 @@
 // src/types.rs
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
@@ -11,27 +12,27 @@ pub enum Side {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Ticker {
     pub symbol: String,
-    pub price: f64,
+    pub price: Decimal, // f64 -> Decimal
     pub timestamp: u64,
 }
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum Signal {
-    Advice(Side, f64),
+    Advice(Side, Decimal), // f64 -> Decimal
     Hold,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Position {
     pub symbol: String,
-    pub quantity: f64,
-    pub entry_price: f64,
-    pub unrealized_pnl: f64,
+    pub quantity: Decimal,       // f64 -> Decimal
+    pub entry_price: Decimal,    // f64 -> Decimal
+    pub unrealized_pnl: Decimal, // f64 -> Decimal
 }
 
 #[derive(Debug, Default, Clone)]
 pub struct Inventory {
-    pub quote_balance: f64,
+    pub quote_balance: Decimal, // f64 -> Decimal
     pub positions: HashMap<String, Position>,
 }
 
@@ -42,7 +43,6 @@ pub struct OrderResponse {
     pub status: String,
 }
 
-// --- Added for TUI ---
 #[derive(Debug, Clone)]
 pub enum UiEvent {
     TickerUpdate(Ticker),
